backend/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
requests>=2.31.0
supabase>=1.0.0
gunicorn>=20.1.0